+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "Remainder fantasy".
+- Tests for "Bust speeding vehicles".
 
 ## [1.16.0] - 2022-12-31
 ### Added